+++ src/main/java/woowacourse/auth/config/AuthenticationPrincipalArgumentResolver.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.ui;
+package woowacourse.auth.config;
 
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
