+++ src/test/java/woowacourse/shoppingcart/ui/CustomerControllerTest.java
@@ -29,7 +29,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import woowacourse.auth.dto.CustomerRequest;
 import woowacourse.auth.dto.CustomerResponse;
-import woowacourse.auth.ui.AuthenticationPrincipalArgumentResolver;
+import woowacourse.auth.config.AuthenticationPrincipalArgumentResolver;
 import woowacourse.shoppingcart.application.CustomerService;
 import woowacourse.shoppingcart.dto.EmailValidationRequest;
 import woowacourse.shoppingcart.exception.InvalidTokenException;
